[ADD] Script to reset the language
This script is useful for implementation time. TYP users are getting
some translation not updated in the instance. To avoid this, it's
necessary to deactivate, remove, and reload the system language. We
don't want to do it every time we update the in-production instance.

I had some +1 in https://github.com/Vauxoo/typ/pull/137

How to Run:

https://asciinema.org/a/cf4wcpw2xdha0uwxy0id6tgjc

Manual solution:

https://drive.google.com/file/d/0B0GA0dsW6w8GbXB3dUx1YVR4VkE/view?usp=sharing