[ADD] odoo-instance: new argument --checkout-commit PATH
When use odoo-instance checkout command now we have an extra
parameter that is --checkout-comm PATH. This let us to pass a path to a
json file that is similar to the config json, but have a extra key per
branch_ commit key telling to odoo-instance in which commit should be
that branch. This command iterate on the branches and apply: git
checkout commit.