[ADD] cloc_multiple_directories: Count the line of code from a repository without the database.
If the directory exists but the database is not available, Cloc documentation does not provide a recursive way to get the number of line of code of every module inside a repository. This script solves that to get a table with each module inside a repository with just one line command and save that table in cloc_file.log.