[IMP] generate_migration_issues.py: add script to get migration status and create issues
Generate a list of Vauxoo modules with a branch in the target Odoo
version and assess their migration status. Process a project
dependencies CSV file, obtained via the
generate_module_dependency_priority.py server action, to determine their
migration status. Create migration issues for the modules that are yet
to be migrated.